Add getServerConfig mwthod to RemoteServerConfiguration
